+++ .gitignore
@@ -9,3 +9,4 @@ out
 gen
 .idea/
 InfectStatistic-example.iml
+/bin/
