+++ .settings/org.eclipse.core.resources.prefs
@@ -1,0 +1,2 @@
+eclipse.preferences.version=1
+encoding//221701322/InfectStatistic.java=UTF-8
