--- 221701322/test
@@ -1,2 +1,0 @@
-this is
-a test
